requirements.txt
Flask==2.0.1
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
wheel
Flask-Cors
pymongo
pymongo[srv]
dnspython
flask-bcrypt
requests
python-dotenv